splitting Slice data
