+++ src/components/Header.js
@@ -1,6 +1,6 @@
 //UNDERSTAND
 import { useSelector, useDispatch } from 'react-redux';
-import { authActions } from '../store';
+import { authActions } from '../store/auth';
 import classes from './Header.module.css';
 
 const Header = () => {
